Examine text improvments to humans and id cards (#3031)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

fixes a bug introduced by guest book that shows "this person is ."
improve id examines and attack self to be a bit clearer
adds a way to quickly figure out the faction of a ship

![image](https://github.com/shiptest-ss13/Shiptest/assets/86381784/40847538-f696-4861-92a8-a3b5a39fd535)

![image](https://github.com/shiptest-ss13/Shiptest/assets/86381784/529d41ce-d142-4893-8d7d-6ce9a1310576)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

bug fix and makes it much easier to read id cards

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
tweak: tweaked id card examine information and formating
fix: human examine bug when middle aged
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->