Tweaks Runner-class and replaces weapons with factory subtype on the Elder (#3059)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

Title. The Runner-class should now be able to dock, and the Elder's
Illestrens + Arrows were replaced with factory variants.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

* Ship working good.
* Elder is an SRM ship, which uses Hunter's Pride weapons. The factory
variants are intended for SRM ships.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
balance: The Elder's Illesterns and Arrows are now factory-variant.
fix: The Runner should be able to dock now.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->