Hardliner Rework Part 2: More Clothing and Roles! (#3098)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

DAMN DANIEL... BACK AT IT AGAIN WITH THE `Pull Requests`:bangbang:

This PR adds new outfits for our next Gorlex splinter group, the
Hardliners!

![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/a9ffc60a-faeb-4e7c-9fc0-39ebb6280083)

This shit is straight up bussin', fr fr.

It also slightly remaps the Komodo to include new Hardliner content, for
more awesome spacetok militia fancams.

Finally, this PR does some slight tweaks with a few outfits from NGR
that turned out to be broken. UwUpsie...
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

https://github.com/shiptest-ss13/Shiptest/assets/77556824/656c82d3-c7ec-4bdd-9846-564c1186fcbf

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl: PositiveEntropy
add: The Hardliners have now made a proper appearance in the Frontier!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Imaginos16 <77556824+Imaginos16@users.noreply.github.com>
Co-authored-by: FalloutFalcon <falloutfalconplays@gmail.com>
Co-authored-by: thgvr <81882910+thgvr@users.noreply.github.com>