refresh:clip part 2.5: pride month speial (#3108)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

![image](https://github.com/shiptest-ss13/Shiptest/assets/58402542/f367f987-e700-46b9-be2c-f9599d07b427)
fannart by Ceceil on discord also i have literally never played this
character once wtf

happy pride mont

this pr adds the outfit for war correspodants, as I was asked to and
would be an intresting job in the future

![image](https://github.com/shiptest-ss13/Shiptest/assets/58402542/4a0a7c7b-82b4-411f-a004-00cb3f6051c1)

im such a fucking sussy baka that the journalists report on my
impostering

:cl:
add: Adds a war correspondent for CLIP.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: rye-rice <58402542+rye-rice@users.noreply.github.com>
Co-authored-by: retlaw34 <58402542+retlaw34@users.noreply.github.com>
Co-authored-by: thgvr <81882910+thgvr@users.noreply.github.com>