Makes motorized wheelchairs consume a reasonable amount of power (#3102)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

Reduces the motorized wheelchair's power usage per move from 100 to 20,
and ups the baseline power efficiency by 1.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Motorized wheelchairs with baseline parts currently consume power on par
with a gygax with it's leg actuators on, and it's still a pretty bad
powerhog even with better parts. The reduced power consumption should
make it actually useable.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
balance: Motorized wheelchairs consume less power on move
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->