Permanently and horribly kills decay generators while fixing actual engines (#3091)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

Radiation collectors were nerfed in an undocumented change in #1900
which reduced the amount of power they could effectively generate to the
point the 1 KW handout the collector provides was making up the majority
of its output
This makes the following changes, which should keep radiation-based
generators effective while removing the effectiveness of decay
generators almost entirely

Radiation pulses now need to have a strength of over 300 to trigger a
radiation collector (this is subtracted from the collector's incoming
power, too), which is conveniently the same amount as produced by the
strongest set of radiation barrels. If you find a decayed supermatter,
go ham I guess
coefficient returned to 100, from 10, which boosts the amount of
produced power from radiation. This offsets the loss of power from the
increased minimum.
power drain rate returned to 4%, from 0.1%, increasing the amount of
power the collectors effectively generate

Current numbers place the twinkleshine's unmodified supermatter at ~40
kW per collector, numbers can be adjusted if necessary

In the event someone wants to spend 5 zillion credits on a supermatter
engine, or the twinkleshine is run again, it might as well produce a
reasonable amount of power

:cl:
balance: radiation collectors are now significantly more stingy about
the potency of the radiation they will process. Don't expect old nuclear
waste or openly stored uranium to do much.
balance: radiation collectors also produce significantly more power from
the remaining viable sources
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Signed-off-by: Theos <theubernyan@gmail.com>