Fixes broken mapping versions of volume pumps for SDMM (#3118)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

fixes path error for mapping versions of volume pumps and correctly
renames volume pump mapping icons

![image](https://github.com/shiptest-ss13/Shiptest/assets/90987989/10cea549-efb6-4e3e-ba60-183ec21d7cad)

![image](https://github.com/shiptest-ss13/Shiptest/assets/90987989/48042aa0-e6d0-482c-8fe5-b7f09119736c)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Makes volume pumps that start on have the on texture in sdmm, and makes
the layer versions correctly show up in sdmm.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
fix: fixed volume pumps for mapping.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->