Gives the lever-action rifle a gunslinger ability and removes the gunslinger damage buff from the shadow. (#3135)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

Gives the lever-action rifle an effect similar to revolver fanning, but
slower and while wielded. ALSO removes the +5 damage and AP

The lever action rifle right now is very underwhelming and i think this
would be a fun way to use it. Also, the +5 damage and AP bonus is stupid
and unneeded now that the .44 roumain round is merged.

:cl:
add: A technique for gunslinger with the lever-action flaming arrow!
Firing it while wielded will quickly and automatically rack the lever,
less quickly than a .38 revolver or fanned shadow though.
del: The +5 damage and AP bonus on the shadow while using it as a
gunslinger.

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->