Partial port of tg points of intrest and improved orbit menu (#3320)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
Squash of #3219 rebased ontop of a revert of #3218 cause i fucked up the
commit history

- [x] Fix admin menu
- [x] Fix auto-observe

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
This massively improves the readability of the orbit menu.
You can now add custom points of interest pretty easily as it uses an
element now.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
add: You can now see ships in the orbit menu and its alot prettier!
code: ported tg points of interest and a much improved orbit menu
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>