valor map fixes (#3440)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
So, currently the Valor is awaiting a remap, but this should fix a few
problems that it currently has, most notably, the large areas, lack of
firelocks and lack of lighting in some rooms.

What's changed:

- Added firelocks to every door, no more mid-corridor firelocks which
don't do much.

- Added firelocks across the cargo-bay door. Similar to the Mudskipper.
- _I understand this may be a contentious change but i'm happy to remove
it if the danger of the holofield falling would still like to be
realised_.

![image](https://github.com/user-attachments/assets/48d83b56-7fec-4fb7-9917-337f7f41d78f)

- Added lighting to the Corpsman prep room, and the EVA Room as it
currently does not spawn with a light at all. (with this, I have also
added extra light switches to said rooms).

- Seperated the large area at the bottom of the ship. At the moment, the
Surgery Room, Morgue Room, Corridor, Corpsman prep, and post-op area all
share the same area. I have given each one a seperate area so you can
control the lighting in each room.

![image](https://github.com/user-attachments/assets/29ed4c5c-d67a-407a-9ff9-587135f92483)

- Similar to the above, I have added seperate areas for the Janitor
Closet, EVA prep, Toilet and Laundry Room as well.

![image](https://github.com/user-attachments/assets/ee47bf9b-b07b-4794-911d-e310cc8db933)



## Why It's Good For The Game
Being able to turn the lights off in the Morgue seperatly from the
entire medbay's is a bit of a no-brainer, and helps enhance the
roleplay. It is a little bit immerssion breaking to go to turn off the
lights in the patient-care area and suddenly you've plunged the surgery
into pitch-darkness.

As for the firelocks, having consistent firelocks that are actually
present on every door, is also a no brainer; but this change brings the
Valor in-line with other IRMG ships in actually having fire-locks on
every door.

## Changelog

:cl:
add: Firelocks to the Valor-Class' Doors
add: Lighting to dark areas on the Valor-Class' Doors
add: New areas on the Valor-Class to seperate rooms
add: Added APC for the Surgical Area
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->