One Shot RPG (#3433)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds a one-shot rocket launcher to the code. Adjusts the code in some
places to support it being truly one shot. Also adds being able to draw
on rocket launchers for mild entertainment.
Rye did the sprite.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
A disposable anti-armor tool is good for people who are poor and or
frontiersmen.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog
Rye, Erika 
:cl:
add: One-Shot Rocket Launcher. A quick solution to an armored problem. 
add: You can now scribble profane things onto the tube of your rocket
launcher. Use this power wisely.
imageadd: Rye: A sprite for a one-shot rocket launcher
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->