Minor Colossus Filter Fix (#3442)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Swaps the positions of the filters on the Colossus so that oxygen is
filtered to the oxygen tank, and nitrogen is filtered into the nitrogen
tank.


![image](https://github.com/user-attachments/assets/253af707-0d9d-4b29-9369-bae5d3ac2c64)

## Why It's Good For The Game

I know that a rework for the Colossus is in progress, but I felt that a
minor fix wouldn't hurt, in the meantime. This prevents pure tanks from
being contaminated with gasses that are not supposed to be there,
messing up the overall gas mix composition of the air.

## Changelog

:cl:
fix: fixed the waste and scrubber gas reclamation filters on the
colossus
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->